Remove `handlers.rs` from rust-cli
