+++ packages/rust-cli/src/actions/mod.rs
@@ -1,0 +1,25 @@
+use crate::events::Event;
+use crate::store::EventStore;
+use crate::{models, projector};
+
+pub mod autobreak;
+pub mod config;
+pub mod edit;
+pub mod favorites;
+pub mod push;
+pub mod start;
+pub mod sync;
+mod utils;
+
+pub fn add_event(
+    tasks: &mut Vec<models::Task>,
+    history: &mut Vec<Event>,
+    event_store: &EventStore,
+    event: Event,
+    success_msg: &str,
+) -> String {
+    event_store.persist(&event);
+    history.push(event);
+    *tasks = projector::restore_state(history);
+    success_msg.to_string()
+}
