+++ packages/rust-cli/src/actions/sync.rs
@@ -1,0 +1,184 @@
+use crate::actions::utils::generate_events_from_server_entries;
+use crate::events::Event;
+use crate::external_models::TaskDto;
+use crate::store::EventStore;
+use crate::{alvtime, external_models, models, projector};
+use chrono::{Datelike, Local, NaiveDate};
+use inquire::Select;
+use std::collections::{HashMap, HashSet};
+use std::time::Instant;
+
+pub fn handle_sync(
+    client: &alvtime::AlvtimeClient,
+    tasks: &mut Vec<models::Task>,
+    history: &mut Vec<Event>,
+    event_store: &EventStore,
+    external_tasks: &[TaskDto],
+) -> String {
+    let current_year = Local::now().year();
+    println!("Fetching time entries for {}...", current_year);
+
+    let mut entries = Vec::new();
+    use std::io::Write;
+
+    for month in 1..=12 {
+        let start_date = NaiveDate::from_ymd_opt(current_year, month, 1).unwrap();
+        let next_month_start = if month == 12 {
+            NaiveDate::from_ymd_opt(current_year + 1, 1, 1).unwrap()
+        } else {
+            NaiveDate::from_ymd_opt(current_year, month + 1, 1).unwrap()
+        };
+        let end_date = next_month_start.pred_opt().unwrap();
+
+        print!(".");
+        std::io::stdout().flush().ok();
+
+        match client.list_time_entries(start_date, end_date) {
+            Ok(mut e) => entries.append(&mut e),
+            Err(e) => return format!("\nFailed to fetch entries: {}", e),
+        };
+    }
+    println!(" Done.");
+
+    // 1. Group server entries by Date
+    let mut entries_by_date: HashMap<NaiveDate, Vec<&external_models::TimeEntryDto>> =
+        HashMap::new();
+    for entry in &entries {
+        if let Ok(date) = NaiveDate::parse_from_str(&entry.date, "%Y-%m-%d") {
+            entries_by_date.entry(date).or_default().push(entry);
+        }
+    }
+    // 2. Identify which dates already have LOCAL events (not generated by sync)
+    let local_dates: HashSet<NaiveDate> = history
+        .iter()
+        .filter_map(|e| {
+            let (date, is_generated) = match e {
+                Event::TaskStarted {
+                    start_time,
+                    is_generated,
+                    ..
+                } => (start_time.date_naive(), *is_generated),
+                Event::BreakStarted {
+                    start_time,
+                    is_generated,
+                } => (start_time.date_naive(), *is_generated),
+                Event::Stopped {
+                    end_time,
+                    is_generated,
+                } => (end_time.date_naive(), *is_generated),
+                Event::Undo { time } => (time.date_naive(), false), // Undo is manual
+                Event::Redo { time } => (time.date_naive(), false), // Redo is manual
+                Event::DayRevised { date, .. } => (*date, false),   // Manual revision
+            };
+            if !is_generated { Some(date) } else { None }
+        })
+        .collect();
+    let mut synced_days = 0;
+
+    // 3. Handle Local Changes (Sync Only)
+    for date in &local_dates {
+        let server_entries = entries_by_date.get(date);
+        let has_server_entries = server_entries.map(|e| !e.is_empty()).unwrap_or(false);
+
+        let day_tasks: Vec<&models::Task> = tasks
+            .iter()
+            .filter(|t| t.start_time.date_naive() == *date && !t.is_break)
+            .collect();
+
+        if day_tasks.is_empty() && !has_server_entries {
+            continue;
+        }
+
+        if has_server_entries {
+            let server_hours: f64 = server_entries.unwrap().iter().map(|e| e.value).sum();
+            let local_hours: f64 = day_tasks
+                .iter()
+                .map(|t| t.duration().num_minutes() as f64 / 60.0)
+                .sum();
+
+            let msg = format!(
+                "Conflict on {}: Local {:.1}h vs Server {:.1}h. What do you want to do?",
+                date, local_hours, server_hours
+            );
+
+            let options = vec!["Discard Local (Sync from Server)", "Skip Day (Keep Local)"];
+
+            match Select::new(&msg, options).prompt() {
+                Ok(choice) => match choice {
+                    "Discard Local (Sync from Server)" => {
+                        // Generate events from server data
+                        let new_events = generate_events_from_server_entries(
+                            *date,
+                            server_entries.unwrap(),
+                            external_tasks,
+                        );
+                        let revised_event = Event::DayRevised {
+                            date: *date,
+                            events: new_events,
+                        };
+                        event_store.persist(&revised_event);
+                        history.push(revised_event);
+                        synced_days += 1;
+                    }
+                    _ => {
+                        // Skip
+                    }
+                },
+                Err(_) => {}
+            }
+        }
+    }
+
+    let mut added_count = 0;
+    let mut dates_to_sync = Vec::new();
+    let mut all_new_events = Vec::new();
+
+    // 4. Process new dates
+    for (date, day_entries) in entries_by_date {
+        if local_dates.contains(&date) {
+            continue; // Skip syncing for days that already have local data (or were handled above)
+        }
+
+        let total_hours: f64 = day_entries.iter().map(|e| e.value).sum();
+        if total_hours == 0.0 {
+            continue;
+        }
+
+        dates_to_sync.push(date);
+        let mut events = generate_events_from_server_entries(date, &day_entries, external_tasks);
+        all_new_events.append(&mut events);
+    }
+
+    if !dates_to_sync.is_empty() {
+        // Explicitly delete existing events for these days from the store
+        event_store.delete_events_for_days(&dates_to_sync);
+
+        // If the date being synced is the current day (today), clear the in-memory history
+        // so we can rebuild it with the fresh events from the server.
+        let today = Local::now().date_naive();
+        if dates_to_sync.contains(&today) {
+            history.clear();
+        }
+
+        event_store.persist_batch(&all_new_events);
+
+        for evt in &all_new_events {
+            // Only push to history if it is for the current day
+            if evt.date() == today {
+                history.push(evt.clone());
+            }
+        }
+        added_count = dates_to_sync.len();
+    }
+
+    if added_count > 0 || synced_days > 0 {
+        // Restore state after bulk insert (re-projects history into tasks)
+        *tasks = projector::restore_state(history);
+        format!(
+            "Synced. Pulled {} days, Overwrote Local {} days.",
+            added_count, synced_days
+        )
+    } else {
+        "No incoming changes.".to_string()
+    }
+}
